Moved state code dictionaries to constants file so the files are more organized and don't hve any extra things they don't need.
