+++ capstone-ui/src/constants.js
@@ -119,4 +119,112 @@ const baseCandidates = {
     "Kanye West": [0, 0]
 }
 
-export { options, unclickables, questionsWithOptions, parties, baseCandidates, partiesOrganized };
+const FIPSfromStateName = {
+  'alabama': '1',
+  'alaska': '2',
+  'arizona': '4',
+  'arkansas': '5',
+  'california': '6',
+  'colorado': '8',
+  'connecticut': '9',
+  'delaware': '10',
+  'district-of-columbia': '11',
+  'florida': '12',
+  'georgia': '13',
+  'hawaii': '15',
+  'idaho': '16',
+  'illinois': '17',
+  'indiana': '18',
+  'iowa': '19',
+  'kansas': '20',
+  'kentucky': '21',
+  'louisiana': '22',
+  'maine': '23',
+  'maryland': '24',
+  'massachusetts': '25',
+  'michigan': '26',
+  'minnesota': '27',
+  'mississippi': '28',
+  'missouri': '29',
+  'montana': '30',
+  'nebraska': '31',
+  'nevada': '32',
+  'new-hampshire': '33',
+  'new-jersey': '34',
+  'new-mexico': '35',
+  'new-york': '36',
+  'north-carolina': '37',
+  'north-dakota': '38',
+  'ohio': '39',
+  'oklahoma': '40',
+  'oregon': '41',
+  'pennsylvania': '42',
+  'rhode-island': '44',
+  'south-carolina': '45',
+  'south-dakota': '46',
+  'tennessee': '47',
+  'texas': '48',
+  'utah': '49',
+  'vermont': '50',
+  'virginia': '51',
+  'washington': '53',
+  'west-virginia': '54',
+  'wisconsin': '55',
+  'wyoming': '56',
+};
+
+const fipsStateCodes = {
+  '1': 'alabama',
+  '2': 'alaska',
+  '4': 'arizona',
+  '5': 'arkansas',
+  '6': 'california',
+  '8': 'colorado',
+  '9': 'connecticut',
+  '10': 'delaware',
+  '11': 'district-of-columbia',
+  '12': 'florida',
+  '13': 'georgia',
+  '15': 'hawaii',
+  '16': 'idaho',
+  '17': 'illinois',
+  '18': 'indiana',
+  '19': 'iowa',
+  '20': 'kansas',
+  '21': 'kentucky',
+  '22': 'louisiana',
+  '23': 'maine',
+  '24': 'maryland',
+  '25': 'massachusetts',
+  '26': 'michigan',
+  '27': 'minnesota',
+  '28': 'mississippi',
+  '29': 'missouri',
+  '30': 'montana',
+  '31': 'nebraska',
+  '32': 'nevada',
+  '33': 'new-hampshire',
+  '34': 'new-jersey',
+  '35': 'new-mexico',
+  '36': 'new-york',
+  '37': 'north-carolina',
+  '38': 'north-dakota',
+  '39': 'ohio',
+  '40': 'oklahoma',
+  '41': 'oregon',
+  '42': 'pennsylvania',
+  '44': 'rhode-island',
+  '45': 'south-carolina',
+  '46': 'south-dakota',
+  '47': 'tennessee',
+  '48': 'texas',
+  '49': 'utah',
+  '50': 'vermont',
+  '51': 'virginia',
+  '53': 'washington',
+  '54': 'west-virginia',
+  '55': 'wisconsin',
+  '56': 'wyoming',
+};
+
+export { options, unclickables, questionsWithOptions, parties, baseCandidates, partiesOrganized, FIPSfromStateName, fipsStateCodes };
